bugfix: ensure no panics when calling t.Name() inside a generator
This is a fairly annoying footgun. While rapid.T is guaranteed to
never be nil pretty much, we offer no such guarantees as to
the presence of testing.T inside of it. As a result
calling APIs present on rapid.T interface can result
in panics if invoked via `.Example()` instead of .Draw()

Attached test case would fail prior to this change and
illustrates the problem